Added LAME notices in Sample project
I think this may not be necessary, since the sample project is not
(and is unlikely to be) distributed separately from this whole
repo, which already acknowledged that LAME was used and included a
URL (as requested in the "LICENSE" file in LAME's source code). I
don't this does any any harm, though.