Remove CodeQL workflow (before reconfiguring CodeQL) (#12)
The current workflow fails to build the project, but it may not
need to do so, since CodeQL does not usually need to build C#
anymore in order to scan it.

There are various other improvements that will come from either
switching to the default/basic setup, or sticking with advanced but
having GitHub regenerate the workflow. This includes some
simplifications, as well as automatically adding GitHub Actions
scanning.